+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.0.2
 
+- Change order so that `(seq-elt set 0)` gives the earliest inserted element
 - Require cl-lib to avoid a `cl-defstruct` not defined error that sometimes pops up
 
 ## 0.0.1